Widen lockUnion to uint48 and make getLockStatus lazily report elapsed unlocks
Normalize the lock timestamp width with expiry: AccountState.lockUnion goes
uint40 -> uint48 (the packed slot still fits in one word; trailing reserved
region shrinks 2 -> 1 byte). getLockStatus/AccountUnlockInitiated widths follow.

getLockStatus now applies the elapse itself: once block.timestamp reaches the
pending unlocksAt it returns the clean unlocked projection (false, false, 0, 0),
matching _checkAndClearLock's post-clear result instead of surfacing the stale
initiated-unlock fields still sitting in storage.

+++ test/unit/Keystore/createAccount.t.sol
@@ -266,7 +266,7 @@ contract CreateAccountTest is KeystoreTest {
         assertEq(keystore.getChangeSequences(account).local, 1);
         assertEq(keystore.getChangeSequences(account).multichain, 0);
 
-        (bool locked, bool hasInitiatedUnlock, uint40 unlocksAt, uint16 unlockDelay) = keystore.getLockStatus(account);
+        (bool locked, bool hasInitiatedUnlock, uint48 unlocksAt, uint16 unlockDelay) = keystore.getLockStatus(account);
         assertFalse(locked);
         assertFalse(hasInitiatedUnlock);
         assertEq(unlocksAt, 0);
